Got all tests working up to 24 processors, and with mixed MPI+OpenMP, hurra.

+++ src/redistribution/folding.h
@@ -67,7 +67,7 @@ namespace CTF_int {
                        int              dir,
                        int              max_ntd,
                        char **          tswap_data,
-                       int *            chunk_size,
+                       int64_t *        chunk_size,
                        algstrct const * sr);
 
 }
